refactor(Button): 음식점 등록 버튼 클릭 시 SelectBox change 이벤트 트리거
